added loggin of operations, add life time confirmation code and unconfirmed user

+++ src/auth/auth.service.ts
@@ -1,6 +1,7 @@
 import {
 	BadRequestException,
 	Injectable,
+	Logger,
 	UnauthorizedException,
 } from '@nestjs/common';
 import { UserService } from 'src/users/users.service';
@@ -8,32 +9,30 @@ import { PasswordService } from './password.service';
 import { JwtService } from '@nestjs/jwt';
 import { MailService } from './mail.service';
 import { UnconfirmedUserService } from 'src/unconfirmedUsers/unconfirmedUsers.service';
+import { ConfirmationCodeService } from './confirmationCode.service';
 
 @Injectable()
 export class AuthService {
+	private readonly logger = new Logger(AuthService.name);
+
 	constructor(
 		private userService: UserService,
 		private unconfirmedUserService: UnconfirmedUserService,
 		private passwordService: PasswordService,
 		private jwtService: JwtService,
-		private mailService: MailService,
+		private confirmationCodeService: ConfirmationCodeService,
+		private mailService: MailService
 	) {}
 
-	async sendConfirmationCodeOnEmail(email: string) {
-		const code = Math.random().toString(36).substring(2, 8).toUpperCase();
-
-		this.mailService.sendConfirmationCode(email, code);
-
-		return code;
-	}
-
 	async updateConfirmationCode(email: string, id: number, isSignUp: boolean) {
-		const code = await this.sendConfirmationCodeOnEmail(email)
+		const {code, codeExpiresAt} = this.confirmationCodeService.getCode()
+
+		await this.mailService.sendConfirmationCode(email, code)
 
 		if(isSignUp) {
-			this.unconfirmedUserService.updateConfirmationCode(id, code)
+			this.unconfirmedUserService.updateConfirmationCode(id, code, codeExpiresAt)
 		} else {
-			this.userService.updateConfirmationCode(id, code)
+			this.userService.updateConfirmationCode(id, code, codeExpiresAt)
 		}
 	}
 
@@ -41,7 +40,7 @@ export class AuthService {
 		const user = await this.userService.findByEmail(email);
 
 		if (user) {
-			throw new BadRequestException(`User with email: ${email} already exists`);
+			throw new BadRequestException(`User with email '${email}' already exists`);
 		}
 
 		const unconfirmedUser =
@@ -53,14 +52,17 @@ export class AuthService {
 
 		const salt = this.passwordService.getSalt();
 		const hash = this.passwordService.getHash(password, salt);
-		const confirmationCode = await this.sendConfirmationCodeOnEmail(email);
+		const {code, codeExpiresAt} = this.confirmationCodeService.getCode()
+
+		await this.mailService.sendConfirmationCode(email, code)
 
 		const newUnconfirmedUser = await this.unconfirmedUserService.create(
 			email,
 			hash,
 			salt,
 			country,
-			confirmationCode,
+			code,
+			codeExpiresAt
 		);
 
 		const accessToken = await this.jwtService.signAsync({
@@ -80,12 +82,14 @@ export class AuthService {
 			throw new BadRequestException('User not found');
 		}
 
-		if (unconfirmedUser.confirmationCode === code) {
+		const isExpired = this.confirmationCodeService.checkOnExpired(unconfirmedUser.confirmationCodeExpiresAt)
+
+		if (!isExpired && unconfirmedUser.confirmationCode === code) {
 			const user = await this.userService.create(
 				unconfirmedUser.email,
 				unconfirmedUser.hash,
 				unconfirmedUser.salt,
-				unconfirmedUser.country,
+				unconfirmedUser.countryCode,
 			);
 
 			await this.unconfirmedUserService.delete(unconfirmedUser.id)
@@ -93,6 +97,7 @@ export class AuthService {
 			const accessToken = await this.jwtService.signAsync({
 				id: user.id,
 				email: user.email,
+				roles: user.roles,
 				isConfirmed: true
 			});
 
@@ -106,13 +111,18 @@ export class AuthService {
 		const user = await this.userService.findById(id)
 
 		if(!user) {
-			throw new UnauthorizedException('The user does not exist');
+			throw new UnauthorizedException('User does not exist');
 		}
 
-		if(user.confirmationCode === code) {
+		const isExpired = this.confirmationCodeService.checkOnExpired(user.confirmationCodeExpiresAt)
+
+		if(!isExpired && user.confirmationCode === code) {
+			await this.userService.updateConfirmationCode(id, null, null)
+
 			const accessToken = await this.jwtService.signAsync({
 				id: user.id,
 				email: user.email,
+				roles: user.roles,
 				isConfirmed: true
 			});
 
@@ -134,6 +144,12 @@ export class AuthService {
 		if (hash !== user.hash) {
 			throw new UnauthorizedException('The user does not exist');
 		}
+
+		const {code, codeExpiresAt} = this.confirmationCodeService.getCode();
+
+		await this.mailService.sendConfirmationCode(email, code)
+
+		await this.userService.updateConfirmationCode(user.id, code, codeExpiresAt)
 
 		const accessToken = await this.jwtService.signAsync({
 			id: user.id,
